feat: minimally working GGRS port
It is not very clean and it doesn't fell very reliable yet, but it works.

It was a struggle to translate some rust stuff like fixed-size arrays
and handle all the mutability involved.

It definitely needs another round of coding and polish:

Better tests. I want to know for sure it is reliable.

Idiomatic interface. Callbacks may be a better fit.

Memory management: I struggled translating some of rust's structures and control flows,
there is very likely some mishandling of memory.